+++ tests/ttnn/unit_tests/gtests/CMakeLists.txt
@@ -63,7 +63,6 @@ target_sources(
         test_graph_query_op_constraints.cpp
         test_graph_query_op_runtime.cpp
         test_launch_operation.cpp # TODO: Fix data race (TSan) then shift-left
-        test_matmul_benchmark.cpp
         test_relational_int.cpp
         test_rsub_int.cpp
         test_sub_int.cpp
